requirements.txt
requests
pandas
numpy
tensorflow
scikit-learn
python-telegram-bot
newsapi-python
vaderSentiment
python-dotenv
